+++ src/views/unprotected/SuperAdminModule/SuperAdminLogin.vue
@@ -121,7 +121,7 @@ const onSubmit = form.handleSubmit(async () => {
           user: { token, firstName, lastName, _id: id, email: userEmail }
         } = data
 
-        const roles: {_id: string, name: string}[] = response.data.data.user.roles
+        const vendors: any[] = response.data.data.user.vendors
 
         // save basic user data to local storage
         setLocalStorage(firstName, lastName, userEmail, id)
@@ -130,14 +130,16 @@ const onSubmit = form.handleSubmit(async () => {
         // Save the token in Pinia store
         setToken(token)
 
-        if(roles){
-          const vendorRole = roles.find((r)=> r.name.toLowerCase() === 'vendor')
-          if(vendorRole){
-            useSuperAdminStore().isVendor = true
-          }
+        if(vendors.length !== 0){
+          useSuperAdminStore().setVendor(true, vendors[0].vendorId)
+          localStorage.setItem('vendor', JSON.stringify(vendors[0].companyName))
+          useSuperAdminStore().vendorId = vendors[0].vendorId
+          useSuperAdminStore().companyName = vendors[0].companyName
+        }else{
+          useSuperAdminStore().setVendor(false, '')
         }
         sessionStorage.setItem('permissions', JSON.stringify(response.data.data.user.permissions))
-        // console.log(response.data)
+        // console.log(response)
         // const permissions = modPermissions(response.data.data.user.permissions)
         // ability.update(permissions)
         // updateAbility(permissions)
@@ -154,6 +156,7 @@ const onSubmit = form.handleSubmit(async () => {
     } catch ({ response }: any) {
       loading.value = false
 
+      // console.log(response)
       const { status, data } = response
 
       if ([400, 401].includes(status)) {
